Move underscore package to underscore-base
