+++ packages/modules/package.js
@@ -6,7 +6,7 @@ Package.describe({
 });
 
 Package.onUse(function(api) {
-  api.use("underscore");
+  api.use("underscore-base");
   api.use("modules-runtime");
   api.mainModule("client.js", "client");
   api.mainModule("server.js", "server");
