+++ packages/underscore-base/README.md
@@ -1,0 +1,8 @@
+# underscore-base
+
+Underscore.js is a popular library that provides various utility functions to
+simplify processing arrays, objects and functions.
+
+To learn more about underscore, see
+[the underscore project site](http://underscorejs.org/).
+
