+++ packages/underscore-base/post.js
@@ -1,3 +1,3 @@
 // This exports object was created in pre.js.  Now copy the `_` object from it
 // into the package-scope variable `_`, which will get exported.
-_ = exports._;
+__ = exports.__;
