--- packages/underscore-tests/README.md
@@ -1,1 +1,0 @@
-This is an internal Meteor package.
