--- packages/underscore-tests/package.js
@@ -1,11 +1,0 @@
-Package.describe({
-  // These tests can't be directly in the underscore packages since
-  // Tinytest depends on underscore
-  summary: "Tests for the underscore package",
-  version: '1.0.6-rc.1'
-});
-
-Package.onTest(function (api) {
-  api.use(['tinytest', 'underscore']);
-  api.addFiles('each_test.js');
-});
